# PERF: Simplify some of loadtxt's standard converters.