@anntzer 2021-08-06
```
$ python runtests.py --bench-compare HEAD~ Loadtxt
<elided>
       before           after         ratio
     [99eac423]       [63e0dc4e]
     <main>           <loadtxtconvstr>
-       170±0.7ms        159±0.3ms     0.94  bench_io.LoadtxtCSVStructured.time_loadtxt_csv_struct_dtype
-      33.4±0.2μs       30.4±0.2μs     0.91  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10)
-      30.5±0.1μs       27.6±0.2μs     0.91  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10)
-         191±1μs        169±0.6μs     0.88  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100)
-      35.9±0.3μs       31.7±0.5μs     0.88  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 10)
-      18.3±0.1ms      16.1±0.09ms     0.88  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10000)
-         182±1ms          161±1ms     0.88  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100000)
-     15.9±0.06ms       13.8±0.2ms     0.87  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10000)
-         169±2μs          146±1μs     0.86  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100)
-       169±0.9ms          145±1ms     0.86  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100000)
-       213±0.9ms        174±0.5ms     0.81  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100000)
-     20.6±0.07ms      16.7±0.09ms     0.81  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 10000)
-         221±2μs        178±0.4μs     0.81  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100)
```